+++ src/math.ts
@@ -982,7 +982,7 @@ export class Mat4 {
             const r = Math.sqrt(this.m[0] * this.m[0] + this.m[1] * this.m[1]);
             return new Vec2(
                 Math.atan(this.m[0] * this.m[4] + this.m[1] * this.m[5])
-                / (r * r),
+                    / (r * r),
                 0,
             );
         } else if (this.m[4] != 0 || this.m[5] != 0) {
@@ -990,7 +990,7 @@ export class Mat4 {
             return new Vec2(
                 0,
                 Math.atan(this.m[0] * this.m[4] + this.m[1] * this.m[5])
-                / (s * s),
+                    / (s * s),
             );
         } else {
             return new Vec2(0, 0);
@@ -1405,7 +1405,7 @@ export function testPolygonPoint(poly: Polygon, pt: Point): boolean {
             ((p[i].y > pt.y) != (p[j].y > pt.y))
             && (pt.x
                 < (p[j].x - p[i].x) * (pt.y - p[i].y) / (p[j].y - p[i].y)
-                + p[i].x)
+                    + p[i].x)
         ) {
             c = !c;
         }
@@ -1423,7 +1423,7 @@ export function testEllipsePoint(ellipse: Ellipse, pt: Point): boolean {
     const vx = pt.x * c + pt.y * s;
     const vy = -pt.x * s + pt.y * c;
     return vx * vx / (ellipse.radiusX * ellipse.radiusX)
-        + vy * vy / (ellipse.radiusY * ellipse.radiusY) < 1;
+            + vy * vy / (ellipse.radiusY * ellipse.radiusY) < 1;
 }
 
 export function testEllipseCircle(ellipse: Ellipse, circle: Circle): boolean {
@@ -2227,7 +2227,7 @@ export class Ellipse {
         const vx = point.x * c + point.y * s;
         const vy = -point.x * s + point.y * c;
         return vx * vx / (this.radiusX * this.radiusX)
-            + vy * vy / (this.radiusY * this.radiusY) < 1;
+                + vy * vy / (this.radiusY * this.radiusY) < 1;
     }
     raycast(origin: Vec2, direction: Vec2): RaycastResult {
         return raycastEllipse(origin, direction, this);
@@ -2281,6 +2281,30 @@ export class Polygon {
     }
 }
 
+export function evaluateQuadratic(
+    pt1: Vec2,
+    pt2: Vec2,
+    pt3: Vec2,
+    t: number,
+) {
+    const t2 = t * t;
+    const mt = 1 - t;
+    const mt2 = mt * mt;
+    return pt1.scale(mt2).add(pt2.scale(2 * mt * t)).add(
+        pt3.scale(t2),
+    );
+}
+
+export function evaluateQuadraticFirstDerivative(
+    pt1: Vec2,
+    pt2: Vec2,
+    pt3: Vec2,
+    t: number,
+) {
+    const mt = 1 - t;
+    return pt2.sub(pt1).add(pt3.sub(pt2)).scale(2);
+}
+
 export function evaluateBezier(
     pt1: Vec2,
     pt2: Vec2,
@@ -2296,6 +2320,49 @@ export function evaluateBezier(
     return pt1.scale(mt3).add(pt2.scale(3 * mt2 * t)).add(
         pt3.scale(3 * mt * t2),
     ).add(pt4.scale(t3));
+}
+
+export function evaluateBezierFirstDerivative(
+    pt1: Vec2,
+    pt2: Vec2,
+    pt3: Vec2,
+    pt4: Vec2,
+    t: number,
+) {
+    const t2 = t * t;
+    const mt = 1 - t;
+    const mt2 = mt * mt;
+    return pt2.sub(pt1).scale(3 * mt2).add(pt3.sub(pt2).scale(6 * mt * t)).add(
+        pt4.sub(pt3).scale(3 * t2),
+    );
+}
+
+export function evaluateCatmullRom(
+    pt1: Vec2,
+    pt2: Vec2,
+    pt3: Vec2,
+    pt4: Vec2,
+    t: number,
+) {
+    const A = 0.5 * (((-t + 2) * t - 1) * t);
+    const B = 0.5 * (((3 * t - 5) * t) * t + 2);
+    const C = 0.5 * (((-3 * t + 4) * t + 1) * t);
+    const D = 0.5 * (((t - 1) * t) * t);
+    return pt1.scale(A).add(pt2.scale(B)).add(pt3.scale(C)).add(pt4.scale(D));
+}
+
+export function evaluateCatmullRomFirstDerivative(
+    pt1: Vec2,
+    pt2: Vec2,
+    pt3: Vec2,
+    pt4: Vec2,
+    t: number,
+) {
+    const A = 0.5 * ((-3 * t + 4) * t - 1);
+    const B = 0.5 * ((9 * t - 10) * t);
+    const C = 0.5 * ((-9 * t + 8) * t + 1);
+    const D = 0.5 * ((3 * t - 2) * t);
+    return pt1.scale(A).add(pt2.scale(B)).add(pt3.scale(C)).add(pt4.scale(D));
 }
 
 export function sat(p1: Polygon, p2: Polygon): Vec2 | null {